Rewriting -> moved to a IONode philosophy
Using a simple IONode as a building block, more generic and complex
network can be used.
In particular, a generic Convolution node (that encompass both
convolution and pooling operations. Which difference stands only on the
type of filter used, its input/output dimensions and the size of the
step size when sliding the window)